requirements.txt
flask
librosa
scikit-learn
numpy
openai-whisper